+++ src/main/java/banter/Banter.java
@@ -1,3 +1,5 @@
+package banter;
+
 public class Banter {
     private Storage storage = new Storage();
     private Parser parser = new Parser(storage);
