+++ src/main/java/banter/Main.java
@@ -1,3 +1,5 @@
+package banter;
+
 public class Main {
     public static void main(String[] args) {
         Banter banter = new Banter();
