+++ src/main/java/banter/enums/CommandType.java
@@ -1,4 +1,4 @@
-package commands;
+package banter.enums;
 
 public enum CommandType {
     BYE, LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE;
