--- src/main/java/errors/InvalidTaskNumberUsageError.java
@@ -1,9 +1,0 @@
-package errors;
-
-import tasks.TaskList;
-
-public class InvalidTaskNumberUsageError extends InvalidBanterUsageError {
-    public InvalidTaskNumberUsageError(String error, String usage, TaskList list) {
-        super(error, usage + "\n" + list);
-    }
-}
